requirements.txt
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-google-genai
python-dotenv
chromadb
sentence-transformers
